requirements.txt
transformers
torch
gradio
Pillow
